FIX: Fixed issue with printing msg.
Always add a null term to recieved data, hence if we are working with a
string we may just print etc. If we are working with anything other than
a string then we must use the len in the struct to chop off the null
term.

This is just a quaility of life feature.