WIP: Threads
Started adding worker functions for threads, need to implement thread
locks to maintain thread safety.

For this we need to have locks for individual queus (locks based on
tags) to allow different tagged messages to be added to thier queues at
the same time.

I also need to implement thread safety for the connected array which
contains information on currently connected sockets.