WIP: Threads Completed
Next I will be adding the ability to toggle the use of the queues, this
will be a nice feature, as it means you can either work with a pure fifo
socket basic socket approach. Or you can activate the queue system and
get fifo based on the tag of a message.

Note when using this library you must now link in pthreads when
compiling.